requirements.txt
# Core dependencies
streamlit==1.31.0
praw==7.7.1
python-dotenv==1.0.0
wordcloud==1.8.1 
# Database
psycopg2-binary==2.9.9
snowflake-connector-python==3.5.0
snowflake-sqlalchemy==1.5.1

# Data processing
pandas==2.1.4
numpy==1.26.4

# Visualization
plotly==5.18.0
matplotlib==3.8.2

# Development
black==24.1.1
flake8==7.0.0
pytest==8.0.0

# Additional dependencies
typing-extensions>=4.0.0
sqlalchemy==1.4.54
alembic>=1.6.0
